define a HeaderTemplate as a header interface, move header implementations into own folder and namespace
